viele funktionale Änderungen und Anpassungen an der GUI
Ordnerstruktur geändert Debug und Release exe aufgeteilt
Installer Ordner hinzugefügt - beinhaltet den Installer, UPX_KiiTree packer und die Installer-Script
---Main_PWM---
Menüleiste hinzugefügt ( speichern, neuer Kii, Kii löschen, Neuer Ordner, Ordner Löschen, A->Z, Z->A )
->AddNewKii - erzeugt einen Kii
->AddNewFolder - erzeugt einen neuen Ordner
->DelKii - löscht einen Schlüssel
->DelFolder - löscht einen Ordner
->SBAbisZ - sortiert die Kii von A bis Z
Checkbox implementiert, für die Auswahl der Themenfarben
Benutzernamen können in die Zwischenablage gespeichert werden
Passwörter können in die Zwischenablage gespeichert werden
implementierung des Masterpasswort ändern Buttons mit Hints
implementierung der autosave Checkbox
VirtualTreeview Zweilenweise anmalen
selber erstellte Ordner können umbenannt werden und bekommen vom user gleich zu beginn einen Namen
Themenfarben können Benutzerspezifisch ausgelwählt und gespeichert werden
KiiTree kann über Strg + S gespeichert werden
in den DBEdit Felder können jetzt markiert Textteile kopiert und eingefügt werden

---Login_PWM---
CheckKTPExistPlus - prüft ob die gewählte datei in den jeweiligen Verzeichnis vorhanden ist, wenn nicht dann wird eine Nachtricht dem User mitgeteilt
implementiertung der USB Geräterkennung, beim einstecken eines USB Geräte wird gefragt, ob das USB Gerät zu speichern vom KiiTree benutz werder soll und passt entsprechenden den Ladepfad an
Drucker implementiert - die eingebenen Passwörter können im Querformat ( der übersichthalber ) ausgedruckt werden
->entsprechende Füllroutine implementiert welche alle Daten sammelt, zurechtrückt und dann an den Drucker schickt

---About_PWM---
Aboutdialog hinzugefügt, beinhaltet allgemeine Infos über den entwickler und dem Programm
passt sich an das gewählte Themen an

---MPW_change_PWM---
neues Formular hinzugefügt zum ändern des Masterpasswortes durch eine 2 Schritte Prüfung
1. Das alte Passwort muss eingegeben werden und mit den jetzigen übereinstimmt
2. Das neue muss 2 mal hintereinander eingeben werden und darf sich nicht von einander unterscheiden
entsprechende Hinweise auf den Buttons hinzugefügt
Dialog hat den gleichen Farbverlauf wie das eingestellt Themen

---Global_PWM---
GetSpecialFolder - gibt den Windowsbenutzer Order "Dokumente" an
es wurde ein globales MainINI Record erzeugt was die Zustände Neuer Benutzer erstellen und lesten Pfad speichert
Speicherroutine für die MainINI implementiert
globale Speicherroutine für die Benutzerdaten implementiert
globale Laderoutine für die Benutzerdaten implementiert
globale Laderoutine für die MainINI implementiert
globale DefaultSetting Record  mit passender initialisierungsroutine implementiert

---PWM_VST---
Sicherheitsabfrage beim löschen eines leeren Ordner hinzugefügt

+++ Units/PWM_VST.pas
@@ -94,9 +94,14 @@ begin
   AVST.Refresh;
   pNode := AVST.GetFirst();
 //  pData := AVST.GetNodeData(PNode);
-  Result := AVST.AddChild( pNode );
+  Result := AVST.InsertNode( pNode,amAddChildLast );
 
   SetData( Result, true, 'Neuer Ordner' );
+
+  AVST.Selected[ Result ] := true;
+  AVST.FocusedNode := Result;
+  AVST.Refresh;
+  Avst.EditNode( Result, -1 );
 
 end;
 
@@ -271,8 +276,14 @@ begin
     end;
   end
   else
-  begin
-    AVST.DeleteNode( pNode );
+  begin  //Change 2020-10-15 sicherheitsabfrage hinzugefügt
+    if MessageDlg( 'Sind Sie sich sicher, dass Sie diesen Ordner löschen wollen?'
+                + sLineBreak + sLineBreak
+                + 'Zu löschender Ordner: "'
+                + pData^.Bezeichnung + '"',
+                mtWarning,
+                [mbYes, mbNo], 0, mbNo ) = 6 then //Ja Btn gedrückt
+      AVST.DeleteNode( pNode );
   end;
 end;
